Refactor integration tests to use ApiFactory and DatabaseFixture; remove ApiTestFixture to streamline test setup.
