--- source/tests/BowlingMegabucks.TournamentManager.IntegrationTests/ApiTestFixture.cs
@@ -1,33 +1,0 @@
-using BowlingMegabucks.TournamentManager.Api;
-using Microsoft.AspNetCore.Mvc.Testing;
-using Microsoft.AspNetCore.Hosting;
-using Testcontainers.MariaDb;
-
-namespace BowlingMegabucks.TournamentManager.IntegrationTests;
-
-public class ApiTestFixture
-    : WebApplicationFactory<IApiMarker>, IAsyncLifetime
-{
-    internal const string _apiKey = "Integration";
-
-    private readonly MariaDbContainer _mariaDbContainer = new MariaDbBuilder()
-        .WithImage("mariadb:11.4.7")
-        .WithDatabase("bowlingmegabucks")
-        .Build();
-
-    protected override void ConfigureWebHost(IWebHostBuilder builder)
-    {
-        Environment.SetEnvironmentVariable("ConnectionStrings:Default", _mariaDbContainer.GetConnectionString());
-        Environment.SetEnvironmentVariable("Authentication:ApiKey", _apiKey);
-    }
-
-    public async ValueTask InitializeAsync()
-        => await _mariaDbContainer.StartAsync();
-
-    public new async Task DisposeAsync()
-    {
-        await _mariaDbContainer.DisposeAsync();
-
-        await base.DisposeAsync();
-    }
-}
