+++ source/tests/BowlingMegabucks.TournamentManager.IntegrationTests/BaseIntegrationTests.cs
@@ -4,14 +4,14 @@ using Microsoft.Extensions.DependencyInjection;
 namespace BowlingMegabucks.TournamentManager.IntegrationTests;
 
 public abstract class BaseIntegrationTests
-    : IClassFixture<ApiTestFixture>
+    : IClassFixture<ApiFactory>
 {
     private readonly IServiceScope _scope;
     internal readonly DataContext _dbContext;
-    
-    protected BaseIntegrationTests(ApiTestFixture fixture)
+
+    protected BaseIntegrationTests(ApiFactory factory)
     {
-        _scope = fixture.Services.CreateScope();
+        _scope = factory.Services.CreateScope();
         _dbContext = _scope.ServiceProvider.GetRequiredService<DataContext>();
     }
 }
